requirements.txt
google-generativeai
python-dotenv
PyPDF2
python-docx
psycopg2-binary
numpy